Extract triangle Count
