+++ lode-ui/src/Model.js
@@ -1,9 +1,10 @@
-import { Box, Flex, Spinner } from "@chakra-ui/react";
+import { Box, Spinner } from "@chakra-ui/react";
 import { OrbitControls } from "@react-three/drei";
 import { Html } from "@react-three/drei";
-import { Canvas, useThree } from "@react-three/fiber";
-import { useEffect, useMemo, useState } from "react";
+import { Canvas } from "@react-three/fiber";
+import { useMemo, useState } from "react";
 import { GLTFLoader } from "three/examples/jsm/loaders/GLTFLoader";
+import TriangleCount from "./TriangleCount";
 
 function Model({ url }) {
   return (
@@ -16,35 +17,6 @@ function Model({ url }) {
         <GltfModel url={url} />
       </Canvas>
     </Box>
-  );
-}
-
-function TriangleCount({ url }) {
-  const { gl } = useThree();
-  const [triangleCount, setTriangleCount] = useState(0);
-  useEffect(() => {
-    setTriangleCount(0);
-    setTimeout(() => {
-      setTriangleCount(gl.info.render.triangles);
-    }, 500);
-    // eslint-disable-next-line
-  }, [url]);
-
-  if (triangleCount === 0) {
-    return null;
-  }
-  return (
-    <Flex
-      as={Html}
-      fullscreen
-      justify="flex-end"
-      direction="column"
-      pointerEvents="none"
-    >
-      <Box bg="rgba(255, 255, 255, 0.5)" p={2}>
-        Triangle count: {triangleCount}
-      </Box>
-    </Flex>
   );
 }
